feat: autoLotto 생성 로직 변경(개별 랜덤 -> shuffle과 인덱싱)
