--- src/main/java/domain/AutoLottoGenerator.java
@@ -1,26 +1,0 @@
-package domain;
-
-import java.util.HashSet;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.stream.Collectors;
-
-public class AutoLottoGenerator implements LottoGenerator {
-
-    private static final int LOTTO_SIZE = 6;
-    private static final int LOTTO_NUMBER_MAX = 45;
-    private static final int LOTTO_NUMBER_UNIT_TO_CORRECT = 1;
-
-    @Override
-    public Lotto generate() {
-        HashSet<LottoNumber> autoLottoNumbers = new HashSet<>();
-        while (autoLottoNumbers.size() < LOTTO_SIZE) {
-            autoLottoNumbers.add(
-                new LottoNumber(
-                    ThreadLocalRandom.current().nextInt(LOTTO_NUMBER_MAX) + LOTTO_NUMBER_UNIT_TO_CORRECT));
-        }
-        return new Lotto(autoLottoNumbers.stream()
-            .sorted()
-            .collect(Collectors.toList()));
-    }
-}
-
