+++ src/main/java/domain/generator/LottoGenerator.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.generator;
+
+import domain.Lotto;
 
 public interface LottoGenerator {
     Lotto generate();
